requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Cors==5.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
gunicorn==23.0.0
ruff==0.9.10
pytest==8.3.5
pytest-flask==1.3.0
